requirements.txt
# Core data manipulation and analysis
pandas>=1.5.0
numpy>=1.24.0

# HTTP requests for API calls
requests>=2.28.0

# Sentiment analysis for news agent
vaderSentiment>=3.3.2

# Plotting and visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Environment management
python-dotenv>=0.19.0

# YAML configuration
pyyaml>=6.0

langgraph==0.6.7

# Testing framework
pytest>=7.0.0